ref: Move test to keep tests together
Move test into groupActivity.spec.jsx
